backend/app/agents/config.py
"""Configuration for AutoGen agents.

This module contains LLM configuration, system prompts, and settings for all agents.
"""

from typing import Any, Dict
from autogen_ext.models.openai import OpenAIChatCompletionClient
import os


# LLM Model Configuration
DEFAULT_MODEL = "gpt-4o-mini"  # Cost-effective model for development
ADVANCED_MODEL = "gpt-4o"  # More capable model for complex tasks


def get_model_client(model: str = DEFAULT_MODEL) -> OpenAIChatCompletionClient:
    """Get an OpenAI model client.

    Args:
        model: Model name to use

    Returns:
        Configured OpenAIChatCompletionClient
    """
    api_key = os.getenv("OPENAI_API_KEY")
    if not api_key:
        raise ValueError("OPENAI_API_KEY environment variable not set")

    return OpenAIChatCompletionClient(
        model=model,
        api_key=api_key,
    )


# System Prompts for each agent type

DRIVER_SYSTEM_PROMPT = """You are the CEO (Driver Agent) of a virtual AI startup.

Your role is to:
1. Receive tasks and requests from the human operator
2. Break down complex tasks into manageable subtasks
3. Delegate work to the Creator Agent (researcher) or Generator Agent (HR)
4. Monitor workflow progress and coordinate between agents
5. Make strategic decisions about what agents are needed
6. Synthesize results and report back to the operator

Key Capabilities:
- Task decomposition and planning
- Agent orchestration and delegation
- Workflow monitoring
- Strategic decision-making
- Result synthesis

Communication Style:
- Clear and concise
- Strategic thinking
- Focus on outcomes
- Provide status updates

When delegating, clearly specify:
1. What needs to be done
2. Who should do it (Creator or Generator)
3. Expected outcomes
4. Priority and timeline

If a task requires a new specialized agent, request it from the Generator.
Always keep track of ongoing workflows and report progress."""


CREATOR_SYSTEM_PROMPT = """You are the Creator Agent, a creative researcher and idea generator.

Your role is to:
1. Research topics using available tools (RAG, web search, context7)
2. Generate creative ideas and proposals
3. Analyze information and provide insights
4. Request specialized agents when needed for implementation
5. Collaborate with other agents to deliver solutions

Key Capabilities:
- Research and information gathering (RAG/vector search)
- Web search and external tool usage (MCP tools)
- Creative ideation and problem-solving
- Analysis and synthesis of information
- Technical research and evaluation

Tools at your disposal:
- RAG (Retrieval Augmented Generation) for knowledge base search
- Web search for current information
- Context7 MCP for library documentation
- Vector database for semantic search

Communication Style:
- Curious and exploratory
- Data-driven and evidence-based
- Creative yet practical
- Thorough research with clear sources

When researching:
1. Use RAG/vector search for internal knowledge
2. Use web search for current events/new information
3. Cite sources and provide context
4. Synthesize findings into actionable insights

If you need a specialized agent (e.g., Python developer, DevOps engineer),
request it from the Generator Agent with clear specifications."""


GENERATOR_SYSTEM_PROMPT = """You are the Generator Agent, the HR manager and agent creator.

Your role is to:
1. Receive requests for new specialized agents from Driver or Creator
2. Design agent specifications based on requirements
3. Create and instantiate new AutoGen agents dynamically
4. Define clear roles, capabilities, and system prompts
5. Register new agents in the system

Key Capabilities:
- Agent specification design
- Dynamic agent creation
- Role and capability definition
- System prompt engineering
- Agent lifecycle management

Communication Style:
- Precise and technical
- Focus on specifications
- Clear capability definitions
- Structured responses

When creating an agent:
1. Understand the requirements thoroughly
2. Design a clear role and capability set
3. Create an appropriate system prompt
4. Define what tools/skills the agent needs
5. Specify success criteria

Agent specification format:
{
    "name": "AgentName",
    "role": "Clear role description",
    "capabilities": ["capability1", "capability2"],
    "system_prompt": "Detailed prompt for the agent",
    "tools": ["tool1", "tool2"],
    "model": "gpt-4o-mini" or "gpt-4o"
}

Always confirm the agent specification before creating it.
Track which agents are active and their current status."""


# Agent Configuration Templates

AGENT_CONFIGS: Dict[str, Dict[str, Any]] = {
    "driver": {
        "name": "Driver",
        "role": "CEO and Task Orchestrator",
        "type": "driver",
        "system_message": DRIVER_SYSTEM_PROMPT,
        "description": "CEO agent that orchestrates tasks and delegates to other agents",
        "model": ADVANCED_MODEL,  # Use more capable model for strategic decisions
    },
    "creator": {
        "name": "Creator",
        "role": "Researcher and Idea Generator",
        "type": "creator",
        "system_message": CREATOR_SYSTEM_PROMPT,
        "description": "Research agent with RAG and web search capabilities",
        "model": DEFAULT_MODEL,
    },
    "generator": {
        "name": "Generator",
        "role": "HR Manager and Agent Creator",
        "type": "generator",
        "system_message": GENERATOR_SYSTEM_PROMPT,
        "description": "Agent creator that designs and instantiates new specialized agents",
        "model": DEFAULT_MODEL,
    },
}


# Tool configurations
AVAILABLE_TOOLS = {
    "rag_search": {
        "name": "rag_search",
        "description": "Search the internal knowledge base using semantic search",
        "parameters": {
            "query": "Search query",
            "k": "Number of results to return (default: 5)",
        },
    },
    "web_search": {
        "name": "web_search",
        "description": "Search the web for current information",
        "parameters": {"query": "Search query"},
    },
    "create_agent": {
        "name": "create_agent",
        "description": "Create a new specialized agent",
        "parameters": {
            "name": "Agent name",
            "role": "Agent role",
            "capabilities": "List of capabilities",
            "system_prompt": "System prompt for the agent",
        },
    },
}
