backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.1.1",
    "flask>=3.1.2",
    "flask-cors>=6.0.1",
    "flask-migrate>=4.1.0",
    "flask-socketio>=5.5.1",
    "flask-sqlalchemy>=3.1.1",
    "pyautogen>=0.10.0",
    "python-dotenv>=1.1.1",
    "python-socketio>=5.14.2",
]

[dependency-groups]
dev = [
    "pyright>=1.1.406",
    "pytest>=8.4.2",
    "pytest-cov>=7.0.0",
    "pytest-mock>=3.15.1",
    "ruff>=0.14.0",
]
